requirements.txt
gradio
requests
google-genai